Moved cards.py into source folder and changed syntax for importing from cards.py
