+++ source/responseCheck.py
@@ -1,4 +1,4 @@
-from cards import handReformat
+from source.cards import handReformat
 
 
 #check if user action matches the correct response
